[Chore #150734699]Folder restructuring
