--- api_endpoints.py
@@ -1,3 +1,0 @@
-from flask_restplus import Resource, Api
-
-api = Api()
