+++ tests/test_api.py
@@ -1,3 +1,4 @@
+"""Module for testing the running of the API"""
 from .basetest import TestBase
